+++ src/components/lyndon/LyndonDesigners.vue
@@ -41,13 +41,13 @@
   </div>
 </template>
 <script>
-import Header from "./layout/Header";
-import Footer from "./layout/Footer";
+import LyndonHeader from "./layout/LyndonHeader";
+import LyndonFooter from "./layout/LyndonFooter";
 export default {
   //   name: Designers
   components: {
-    "app-header": Header,
-    "app-footer": Footer
+    "app-header": LyndonHeader,
+    "app-footer": LyndonFooter
   },
   data() {
     return {
