+++ src/components/lyndon/LyndonProjectDetails.vue
@@ -64,14 +64,14 @@
   </div>
 </template>
 <script>
-import Header from "./layout/Header";
-import Footer from "./layout/Footer";
+import LyndonHeader from "./layout/LyndonHeader";
+import LyndonFooter from "./layout/LyndonFooter";
 import moment from "moment";
 export default {
   //   name: Designers
   components: {
-    "app-header": Header,
-    "app-footer": Footer
+    "app-header": LyndonHeader,
+    "app-footer": LyndonFooter
   },
   data() {
     return {
